Maint: Delete downloads that aren't part of the package
Previously, if the downloads directory contained a file/dir that wasn't
a repo we wanted to package in the installer, e.g. ruby.zip left over
from the old way of packaging ruby, then the tasks would copy it to the
stagedir, and then run candle on it, causing errors.

This commit adds a task to walk each file/dir within downloads
directory, and remove any that are no repos we want to include in the
package.