Update service names to match PE Services when enterprise
This commit chooses the service name for puppet dynamically based upon whether
the build is enterprise or not. If enterprise, the puppet service is named
'pe-puppet' to match our other platforms. Otherwise it is named 'puppet'. As
mcollective is only an enterprise feature, its service name is always
'pe-mcollective' to match our enterprise platform's mco service name.