(PUP-3003) Move event message dll out of ext directory
Previously, the MSI would install the event message dll in
puppet\ext\windows\eventlog\puppetres.dll relative to the installation
directory, and would create a registry value that pointed to the
dll. This is a problem, because we don't want to bundle the ext
directory in the installer.

This commit changes the build process so that the puppetres.dll is
copied into the misc directory (relative to the installation
directory), and updates the registry value to point to the new
location.