(PUP-3003) Explicitly copy service code into staging directory
During WIX packaging, the puppet and mcollective services were
magically copied from the staging\puppet\ext\windows\service and
staging\mcollective\ext\windows directories to the service directory
relative to the installation directory as a result of the File
elements appearing within the Directory element named 'service'.

This is a problem because we don't want to bundle the ext directory
into the installer, and it's not readily apparent how the files end up
in the services directory.

This commit adds a `service` task, which copies the service files from
downloads to staging/service. We continue to rename mcollective's
daemon.bat to mco_daemon.bat, so as to not conflict with puppet's
daemon.bat. The puppet.wxs file then references the locations in
staging/service.